CHAD-17086: zwave-electric-meter lazy loading of sub-drivers
