+++ drivers/SmartThings/zwave-electric-meter/src/aeon-meter/fingerprints.lua
@@ -1,0 +1,9 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local AEON_FINGERPRINTS = {
+  {mfr = 0x0086, prod = 0x0002, model = 0x0009},  -- DSB09xxx-ZWUS
+  {mfr = 0x0086, prod = 0x0002, model = 0x0001},  -- DSB28-ZWEU
+}
+
+return AEON_FINGERPRINTS
