+++ drivers/SmartThings/zwave-electric-meter/src/aeotec-gen5-meter/fingerprints.lua
@@ -1,0 +1,9 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local AEOTEC_GEN5_FINGERPRINTS = {
+  {mfr = 0x0086, prod = 0x0102, model = 0x005F},  -- Aeotec Home Energy Meter (Gen5) US
+  {mfr = 0x0086, prod = 0x0002, model = 0x005F},  -- Aeotec Home Energy Meter (Gen5) EU
+}
+
+return AEOTEC_GEN5_FINGERPRINTS
