+++ drivers/SmartThings/zwave-electric-meter/src/qubino-meter/fingerprints.lua
@@ -1,0 +1,9 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local QUBINO_FINGERPRINTS = {
+  {mfr = 0x0159, prod = 0x0007, model = 0x0054},  -- Qubino 3 Phase Meter
+  {mfr = 0x0159, prod = 0x0007, model = 0x0052}   -- Qubino Smart Meter
+}
+
+return QUBINO_FINGERPRINTS
