+++ drivers/SmartThings/zwave-electric-meter/src/qubino-meter/init.lua
@@ -1,16 +1,6 @@
--- Copyright 2022 SmartThings
---
--- Licensed under the Apache License, Version 2.0 (the "License");
--- you may not use this file except in compliance with the License.
--- You may obtain a copy of the License at
---
---     http://www.apache.org/licenses/LICENSE-2.0
---
--- Unless required by applicable law or agreed to in writing, software
--- distributed under the License is distributed on an "AS IS" BASIS,
--- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
--- See the License for the specific language governing permissions and
--- limitations under the License.
+-- Copyright 2022 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
 
 local capabilities = require "st.capabilities"
 --- @type st.zwave.CommandClass.Configuration
@@ -20,22 +10,10 @@ local Meter = (require "st.zwave.CommandClass.Meter")({ version=3 })
 --- @type st.zwave.CommandClass
 local cc = require "st.zwave.CommandClass"
 
-local QUBINO_FINGERPRINTS = {
-  {mfr = 0x0159, prod = 0x0007, model = 0x0054},  -- Qubino 3 Phase Meter
-  {mfr = 0x0159, prod = 0x0007, model = 0x0052}   -- Qubino Smart Meter
-}
 
 local POWER_UNIT_WATT = "W"
 local ENERGY_UNIT_KWH = "kWh"
 
-local function can_handle_qubino_meter(opts, driver, device, ...)
-  for _, fingerprint in ipairs(QUBINO_FINGERPRINTS) do
-    if device:id_match(fingerprint.mfr, fingerprint.prod, fingerprint.model) then
-      return true
-    end
-  end
-  return false
-end
 
 local function meter_report_handler(self, device, cmd)
 
@@ -101,7 +79,7 @@ local qubino_meter = {
     init = device_init
   },
   NAME = "qubino meter",
-  can_handle = can_handle_qubino_meter
+  can_handle = require("qubino-meter.can_handle"),
 }
 
 return qubino_meter
