+++ drivers/SmartThings/zwave-electric-meter/src/sub_drivers.lua
@@ -1,0 +1,10 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local lazy_load_if_possible = require "lazy_load_subdriver"
+local sub_drivers = {
+   lazy_load_if_possible("qubino-meter"),
+   lazy_load_if_possible("aeotec-gen5-meter"),
+   lazy_load_if_possible("aeon-meter"),
+}
+return sub_drivers
